k8s/watchers/subscriber: Disambiguate API by object type
The initial implementation of this package assumed that Golang allowed
methods of the same name but differing signatures to be defined. That is
not the case. For example, the following code doesn't compile:

func (o *object) OnUpdate(t T) error {}
func (o *object) OnUpdate(u U) error {}

To ease the users of the subscriber package, this commit disambiguates
the various *Handler interfaces. Now, *Handler interface implementations
must have methods with the object name specified, such as
OnUpdateService(), for the ServiceHandler. The reason is to avoid a
situation where a single object that implements multiple `*Handler`s,
for example to react to K8s service and node events, can collide.

Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>